// src/ts/models/button-colors.ts
export interface ButtonColors {
    default: {
        backgroundColor: string;
        color: string;
        border: string;
    },
    onExit: {
        backgroundColor: string;
        color: string;
        border: string;
    },
    onHover: {
        backgroundColor: string;
        color: string;
        border: string;
    },
    onMouseDown: {
        backgroundColor: string;
        color: string;
        border: string;
    },
    onMouseUp: {
        backgroundColor: string;
        color: string;
        border: string;
    }
}